fix: fixed some warnings

+++ include/logger.hpp
@@ -31,7 +31,7 @@ void logMsg(LogLabel level, const std::string &msg,
 class FileLogger {
 private:
   const int BUFFER_SIZE = 16 * 1024;
-  const int MAX_FILE_SIZE = 100 * 1024;
+  const size_t MAX_FILE_SIZE = 100 * 1024;
   const std::string FILE_NAME = std::string(getenv("HOME")) + "/.local/state/wallflower/wallflower.log";
   std::fstream file;
   std::vector<char> buffer;
